refactor: 채팅 모달창과 내부 컴포넌트 분리
